+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 Change Log
 ==========
 
+## Version 1.1.2
+
+_2016-07-19_
+
+* Now it's possible to change the Search View input type.
+* Added the `getItemAtPosition()` method to simplify how you get the list item String.
+
+**e.g.:** `search_view.setInputType(InputType.TYPE_CLASS_TEXT);`
+
 ## Version 1.1.1
 
 _2016-06-23_